+++ app/src/main/AndroidManifest.xml
@@ -2,20 +2,17 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="org.sopt.androidseminar">
 
-    <uses-permission android:name="android.permission.INTERNET" />
-
     <application
-        android:usesCleartextTraffic="true"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.AndroidSeminar">
-        <activity android:name=".presentation.home.UserInfoActivity"/>
-        <activity android:name=".presentation.home.HomeActivity"/>
-        <activity android:name=".presentation.signup.SignUpActivity" />
-        <activity android:name=".presentation.signin.SignInActivity">
+        <activity android:name=".UserInfoActivity"/>
+        <activity android:name=".HomeActivity"/>
+        <activity android:name=".SignupActivity" />
+        <activity android:name=".SignInActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
