+++ app/src/main/java/org/sopt/androidseminar/FollowingUserInfo.kt
@@ -1,0 +1,6 @@
+package org.sopt.androidseminar
+
+data class FollowingUserInfo(
+    val userImage: String,
+    val userName: String
+)
