+++ app/src/main/java/org/sopt/androidseminar/HomeActivity.kt
@@ -1,0 +1,19 @@
+package org.sopt.androidseminar
+
+import android.content.Intent
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import org.sopt.androidseminar.databinding.ActivityHomeBinding
+import org.sopt.androidseminar.databinding.ActivitySignupBinding
+
+class HomeActivity : AppCompatActivity() {
+
+    private lateinit var binding: ActivityHomeBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivityHomeBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+    }
+
+}
