--- app/src/main/java/org/sopt/androidseminar/api/ServiceCreator.kt
@@ -1,14 +1,0 @@
-package org.sopt.androidseminar.api
-
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-//구현체
-object ServiceCreator {
-    private const val BASE_URL = "http://cherishserver.com"
-
-    private val retrofit: Retrofit = Retrofit.Builder()
-        .baseUrl(BASE_URL)
-        .addConverterFactory(GsonConverterFactory.create())
-        .build()
-    val soptService: SoptService = retrofit.create(SoptService::class.java)
-}
