--- app/src/main/java/org/sopt/androidseminar/data/FollowingUserInfo.kt
@@ -1,6 +1,0 @@
-package org.sopt.androidseminar.data
-
-data class FollowingUserInfo(
-        val userImage: String,  //인터넷에서 가져올 예정~
-        val userName: String
-)
