--- app/src/main/java/org/sopt/androidseminar/data/GithubReposInfo.kt
@@ -1,7 +1,0 @@
-package org.sopt.androidseminar.data
-
-data class GithubReposInfo (
-    val reposName: String,
-    val aboutRepos: String,
-    val language: String
-)
