--- app/src/main/java/org/sopt/androidseminar/data/request/RequestLoginData.kt
@@ -1,9 +1,0 @@
-package org.sopt.androidseminar.data.request
-
-import com.google.gson.annotations.SerializedName
-
-data class RequestLoginData(
-    @SerializedName("email")
-    val id: String,
-    val password: String
-)
