--- app/src/main/java/org/sopt/androidseminar/data/request/RequestSignUpData.kt
@@ -1,13 +1,0 @@
-package org.sopt.androidseminar.data.request
-
-import com.google.gson.annotations.SerializedName
-
-data class RequestSignUpData (
-    @SerializedName("email")
-    val id: String,
-    val password: String,
-    val sex: String,
-    val nickname: String,
-    val phone: String,
-    val birth: String
-)
