--- app/src/main/java/org/sopt/androidseminar/data/response/ResponseLoginData.kt
@@ -1,16 +1,0 @@
-package org.sopt.androidseminar.data.response
-
-import com.google.gson.annotations.SerializedName
-
-data class ResponseLoginData (
-    val success: Boolean,
-    val message: String,
-    val data: LoginData?
-)
-
-data class LoginData(
-    @SerializedName("UserId")
-    val userId: Int,
-    val user_nickname: String,
-    val token: String
-)
