--- app/src/main/res/drawable/input_focus_false.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-    <corners android:radius="5dp" />
-    <solid android:color="@color/grayF2" />
-</shape>
