--- app/src/main/res/drawable/input_focus_selector.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:state_focused="true"
-        android:drawable="@drawable/input_focus_true" />
-    <item android:state_focused="false"
-        android:drawable="@drawable/input_focus_false" />
-</selector>
