--- app/src/main/res/drawable/input_focus_true.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-    <corners android:radius="5dp"/>
-    <solid android:color="@color/white"/>
-    <stroke android:width="1dp"
-        android:color="@color/sopt_color"/>
-
-</shape>
