--- app/src/main/res/drawable/item_box.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-    <stroke android:width="1.5dp"
-        android:color="@color/grayF2"/>
-
-</shape>
