+++ app/src/main/res/layout/activity_home.xml
@@ -4,82 +4,50 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.home.HomeActivity">
+    tools:context=".HomeActivity">
 
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/cl_home"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:paddingHorizontal="20dp"
-        android:paddingVertical="15dp"
-        android:background="@color/sopt_color">
-
-        <de.hdodenhof.circleimageview.CircleImageView
-            xmlns:app="http://schemas.android.com/apk/res-auto"
-            android:id="@+id/img"
-            android:layout_width="100dp"
-            android:layout_height="0dp"
-            android:src="@drawable/img"
-            app:layout_constraintDimensionRatio="W, 1:1"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            android:layout_marginTop="20dp"/>
-
-
-        <TextView
-            android:id="@+id/tv_github_id"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="lydiacho"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/img"
-            android:textColor="@color/white" />
-
-        <TextView
-            android:id="@+id/tv_name"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="SeungHee Cho"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/tv_github_id"
-            android:textStyle="bold"
-            android:textColor="@color/white"
-            android:textSize="17sp"/>
-
-        <TextView
-            android:id="@+id/tv_intro"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:text="안알못에서 안린이로 가는중...loading..."
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/tv_name"
-            android:textColor="@color/white"/>
-
-        <Button
-            android:id="@+id/btn_more"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            app:layout_constraintBottom_toBottomOf="@+id/tv_intro"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toTopOf="@id/tv_name"
-            android:text="MORE"
-            android:textColor="@color/sopt_color"
-            android:backgroundTint="@color/white"/>
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rv_repos_list"
-        android:layout_width="match_parent"
+    <ImageView
+        android:id="@+id/img"
+        android:layout_width="0dp"
         android:layout_height="0dp"
-        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-        app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/cl_home"
-        app:layout_constraintVertical_bias="1.0"
-        tools:listitem="@layout/item_repos" />
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        android:layout_marginHorizontal="150dp"
+        android:layout_marginTop="50dp"
+        app:layout_constraintDimensionRatio="1:1"
+        android:src = "@drawable/img" />
 
+    <TextView
+        android:id="@+id/githubID"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="lydiacho"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/img"
+        android:layout_marginTop="10dp"
+        android:layout_marginStart="30dp"
+        android:textSize="10sp" />
+
+    <TextView
+        android:id="@+id/name"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="SeungHee Cho"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/githubID"
+        android:layout_marginStart="30dp"
+        android:textSize="20sp"
+        android:textStyle="bold"/>
+
+    <TextView
+        android:id="@+id/introduction"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:text="자기소개 채워주세유"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/name"
+        android:layout_marginStart="30dp"
+        android:layout_marginHorizontal="30dp"/>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
