+++ app/src/main/res/layout/activity_sign_in.xml
@@ -4,72 +4,64 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:paddingHorizontal="20dp"
-    tools:context=".presentation.signin.SignInActivity">
-
-    <ImageView
-        android:id="@+id/iv_github"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:src="@drawable/img_github"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        android:layout_marginTop="40dp" />
+    tools:context=".SignInActivity">
 
     <TextView
-        android:id="@+id/tv_id"
+        android:id="@+id/ID"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="깃허브 아이디"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        android:layout_marginTop="150dp" />
+        android:layout_marginTop="150dp"
+        android:layout_marginStart="30dp" />
 
     <EditText
-        android:id="@+id/et_id"
-        android:layout_width="match_parent"
+        android:id="@+id/editID"
+        android:layout_width="0dp"
         android:layout_height="wrap_content"
-        app:layout_constraintTop_toBottomOf="@id/tv_id"
-        android:paddingVertical="13dp"
-        android:paddingStart="16dp"
-        android:background="@drawable/input_focus_selector"
+        app:layout_constraintStart_toStartOf="@id/ID"
+        app:layout_constraintTop_toBottomOf="@id/ID"
+        app:layout_constraintEnd_toEndOf="parent"
+        android:layout_marginEnd="30dp"
         android:hint="ID를 입력해주세요"
         android:textSize="15sp" />
 
     <TextView
-        android:id="@+id/tv_pw"
+        android:id="@+id/PW"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="비밀번호"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/et_id"
-        android:layout_marginTop="20dp" />
+        app:layout_constraintTop_toBottomOf="@id/editID"
+        android:layout_marginTop="20dp"
+        android:layout_marginStart="30dp" />
 
     <EditText
-        android:id="@+id/et_pw"
-        android:layout_width="match_parent"
+        android:id="@+id/editPW"
+        android:layout_width="0dp"
         android:layout_height="wrap_content"
-        app:layout_constraintTop_toBottomOf="@id/tv_pw"
-        android:paddingVertical="13dp"
-        android:paddingStart="16dp"
+        app:layout_constraintStart_toStartOf="@id/PW"
+        app:layout_constraintTop_toBottomOf="@id/PW"
+        app:layout_constraintEnd_toEndOf="parent"
+        android:layout_marginEnd="30dp"
         android:hint="비밀번호를 입력해주세요"
         android:inputType="textPassword"
-        android:background="@drawable/input_focus_selector"
         android:textSize="15sp" />
 
     <Button
-        android:id="@+id/btn_login"
-        android:layout_width="match_parent"
+        android:id="@+id/loginButton"
+        android:layout_width="0dp"
         android:layout_height="wrap_content"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         android:layout_marginTop="350dp"
-        android:text="LOGIN"
-        android:textColor="@color/white"
-        android:backgroundTint="@color/sopt_color"/>
+        android:layout_marginHorizontal="30dp"
+        android:text="LOGIN" />
 
     <TextView
-        android:id="@+id/tv_no_id"
+        android:id="@+id/noID"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="아이디/비밀번호가 없으신가요?"
@@ -79,13 +71,13 @@
         android:layout_marginTop="450dp"
         android:layout_marginStart="70dp" />
     <TextView
-        android:id="@+id/tv_sign_up_btn"
+        android:id="@+id/signButton"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="회원가입"
         android:textStyle="bold"
-        app:layout_constraintStart_toEndOf="@id/tv_no_id"
-        app:layout_constraintTop_toTopOf="@id/tv_no_id"
+        app:layout_constraintStart_toEndOf="@id/noID"
+        app:layout_constraintTop_toTopOf="@id/noID"
         android:layout_marginStart="20dp" />
