+++ app/src/main/res/values/colors.xml
@@ -7,6 +7,4 @@
     <color name="teal_700">#FF018786</color>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
-    <color name="grayF2">#F2F2F2</color>
-    <color name="sopt_color">#4260FF</color>
 </resources>
